backend/requirements.txt
alembic==1.7.4
astroid==2.8.4
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cloudinary==1.28.0
coverage==6.3.1
defusedxml==0.7.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
flake8==4.0.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-serialize==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
httpie==2.6.0
idna==3.3
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
mccabe==0.6.1
packaging==21.3
permissive-dict==1.0.4
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.11.1
pyparsing==3.0.7
PySocks==1.7.1
pytest==8.0.0
python-dotenv==0.19.1
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.26
SQLAlchemy-serializer==1.4.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
visitor==0.1.3
Werkzeug==2.0.2
wrapt==1.13.2
WTForms==2.3.3
